requirements.txt
django
python-decouple
djangorestframework
djangorestframework-simplejwt